Conflicts:
	src/applicationContext.xml

+++ src/com/dms/controller/OrderController.java
@@ -4,7 +4,10 @@ import java.io.IOException;
 import java.util.Date;
 import java.util.List;
 
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import org.apache.ibatis.annotations.Param;
 import org.springframework.context.ApplicationContext;
@@ -14,6 +17,7 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.servlet.ModelAndView;
 
 import com.dms.dao.impl.PartDaoImpl;
 import com.dms.entity.CarType;
@@ -28,15 +32,86 @@ import com.dms.service.impl.CarTypeServiceImpl;
 
 @Controller
 public class OrderController {
-	//业务维修开单的空白页
+	
+	/**
+	 * 业务维修开单的空白页
+	 * @param model
+	 * @param request
+	 * @param response
+	 * @throws IOException
+	 * @throws ServletException
+	 */
 	@RequestMapping("order.do")
-	public String order(Model model) {
-		return "orderIndex";
+	public void order(Model model,HttpServletRequest request,HttpServletResponse response) throws IOException, ServletException {
+		//下面验证是否登录，登录正常跳转，否则跳转登录页
+		HttpSession session = request.getSession(false);
+		if(session==null){
+			//转发到登陆页
+			response.sendRedirect(request.getContextPath()+"/login.jsp");
+		} else {
+			String name = (String) session.getAttribute("userName");
+			if(name==null) {
+				//有session，没有我们自己设置的name就说明没有登录，转发到登陆页
+				response.sendRedirect(request.getContextPath()+"/login.jsp");
+			}else {
+				//取出session权限，判断是否具有开单权限。以异步方式，返回数据给ajax的回调函数
+				List<Integer> powerIds = (List<Integer>) session.getAttribute("powerIds");
+				//1在数据库的权限表就代表维修开单
+				if(powerIds.contains(1)){
+					response.getWriter().write("{\"valid\":1,\"url\":\""+request.getContextPath()+"/orderIndex.do\"}");
+				} else {
+					response.getWriter().write("{\"valid\":0,\"message\":\"对不起，没有开单权限\"}");
+				}
+			}
+				
+		}
+
 	}
+	
+	
+	/**
+	 * 再上面方法验证具有开单权限后，会重定向到这，再由这转发到开单页面。
+	 * @param model
+	 * @param request
+	 * @param response
+	 * @return
+	 * @throws IOException
+	 */
+	@RequestMapping("orderIndex.do")
+	public String orderIndex(Model model,HttpServletRequest request,HttpServletResponse response) throws IOException {
+		//下面验证是否登录，登录正常跳转，否则跳转登录页
+		HttpSession session = request.getSession(false);
+		if(session==null){
+			return "redirect:/login.jsp";
+		} else {
+			String name = (String) session.getAttribute("userName");
+			if(name==null) {
+				return "redirect:/login.jsp";
+			}
+		}
+		
+		//要转发前，再次判断是否数据权限， 避免没有权限的账户登录后，  直接输入 这的url进入到这
+		List<Integer> powerIds = (List<Integer>) session.getAttribute("powerIds");
+		if(powerIds.contains(1)){
+			return "orderIndex";
+		}else{
+			return "redirect:/login.jsp";
+		}
+		
+	}
+	
+	
+	
+	
+	
+	
+	
 	
 	//订单中，搜索材料的显示详情
 	@RequestMapping("partList.do")
 	public String partList(Model model, @RequestParam(value = "partname") String partname, @RequestParam(value = "partno") String partno, @RequestParam(value = "modelid") String modelid) {
+		
+		
 		Integer modelId = 0;
 		if(modelid!=null) {
 			modelId= Integer.parseInt(modelid);
@@ -57,7 +132,18 @@ public class OrderController {
 	}
 	//订单中选择添加材料
 	@RequestMapping("addPart.do")
-	public String addpart(Model model) {
+	public String addpart(Model model,HttpServletRequest request) {
+		//下面验证是否登录，登录正常跳转，否则跳转登录页
+		HttpSession session = request.getSession(false);
+		if(session==null){
+			return "redirect:/login.jsp";
+		} else {
+			String name = (String) session.getAttribute("userName");
+			if(name==null) {
+				return "redirect:/login.jsp";
+			}
+		}
+		
 		CarTypeService service = new CarTypeServiceImpl();
 		List<CarType> carTypeList = service.getAllCarType();
 		model.addAttribute("carTypes", carTypeList);
@@ -66,7 +152,19 @@ public class OrderController {
 	}
 	//整单派工时查询所有工人
 	@RequestMapping("allWorkSelect.do")
-	public String allWorkSelect(Model model) {
+	public String allWorkSelect(Model model, HttpServletRequest request) {
+		//下面验证是否登录，登录正常跳转，否则跳转登录页
+		HttpSession session = request.getSession(false);
+		if(session==null){
+			return "redirect:/login.jsp";
+		} else {
+			String name = (String) session.getAttribute("userName");
+			if(name==null) {
+				return "redirect:/login.jsp";
+			}
+		}
+		
+		
 		ApplicationContext ctx =new ClassPathXmlApplicationContext("applicationContext.xml");
 		OrderService  service = (OrderService)ctx.getBean("orderServiceImpl");
 		List<Employee> Ma_Tec = service.getMa_Tec();
@@ -77,7 +175,18 @@ public class OrderController {
 	
 	//根据车牌或者VIN 获取到车辆客户信息，返回给页面
 	@RequestMapping("orderInfo.do")
-	public String getCustomerCarInfoByPlateNumber(Model model, @RequestParam(value = "plateNumber") String plateNumber, @RequestParam(value = "VIN") String VIN) {
+	public String getCustomerCarInfoByPlateNumber(Model model, @RequestParam(value = "plateNumber") String plateNumber, @RequestParam(value = "VIN") String VIN, HttpServletRequest request) {
+		//下面验证是否登录，登录正常跳转，否则跳转登录页
+		HttpSession session = request.getSession(false);
+		if(session==null){
+			return "redirect:/login.jsp";
+		} else {
+			String name = (String) session.getAttribute("userName");
+			if(name==null) {
+				return "redirect:/login.jsp";
+			}
+		}
+		
 		
 		ApplicationContext ctx =new ClassPathXmlApplicationContext("applicationContext.xml");
 		OrderService  service = (OrderService)ctx.getBean("orderServiceImpl");
@@ -101,9 +210,27 @@ public class OrderController {
 	}
 	
 	// 保存新增维修单
-	@ResponseBody
 	@RequestMapping("orderSave.do")
-	public String orderSave(Model model, Order order) {
+	public void orderSave(Model model, Order order, HttpServletRequest request, HttpServletResponse response) throws IOException {
+		//下面验证是否登录，登录正常跳转，否则跳转登录页
+		HttpSession session = request.getSession(false);
+		if(session==null){
+			response.sendRedirect(request.getContextPath()+"/login.jsp");
+			return;
+		} else {
+			String name = (String) session.getAttribute("userName");
+			if(name==null) {
+				response.sendRedirect(request.getContextPath()+"/login.jsp");
+				return;
+			}
+		}
+		
+		//如果保存的数据无效，直接跳转到开单页面。  当然开单页面会验证是否有权限
+		if(order.getCustomerCarInfo()==null||order.getMileage()==null||order.getSenderPhone()==null){
+			response.sendRedirect(request.getContextPath()+"/orderIndex.do");
+			return;
+		}
+		
 		ApplicationContext ctx =new ClassPathXmlApplicationContext("applicationContext.xml");
 		OrderService  service = (OrderService)ctx.getBean("orderServiceImpl");
 		Order saveOrder = service.saveOrder(order);
@@ -111,18 +238,41 @@ public class OrderController {
 		String orderId = saveOrder.getOrderId();
 		String date = saveOrder.getDate();
 		
-		return "{\"date\":\""+ date +"\",\"orderId\":\""+ orderId +"\"}";
+		response.getWriter().print("{\"date\":\""+ date +"\",\"orderId\":\""+ orderId +"\"}");
 	}
 	
+	
+	
 	//作废维修单
-	@ResponseBody
 	@RequestMapping("orderDelete.do")
-	public String orderDelete(HttpServletResponse response,Model model, @RequestParam(value = "orderId") String orderId) throws IOException {
-		ApplicationContext ctx =new ClassPathXmlApplicationContext("applicationContext.xml");
-		OrderService  service = (OrderService)ctx.getBean("orderServiceImpl");
-		service.deleteOrder(orderId);
-        return "{\"data\":\"成功\"}";
+	public void orderDelete(HttpServletRequest request, HttpServletResponse response, Model model, @RequestParam(value = "orderId") String orderId) throws IOException {
+		//下面验证是否登录，登录正常跳转，否则跳转登录页
+		HttpSession session = request.getSession(false);
+		if(session==null){
+			response.sendRedirect(request.getContextPath()+"/login.jsp");
+			return;
+		} else {
+			String name = (String) session.getAttribute("userName");
+			if(name==null) {
+				response.sendRedirect(request.getContextPath()+"/login.jsp");
+				return;
+			}
+		}
+		
+		List<Integer> powerIds = (List<Integer>) session.getAttribute("powerIds");
+		if(powerIds.contains(2)){
+			ApplicationContext ctx =new ClassPathXmlApplicationContext("applicationContext.xml");
+			OrderService  service = (OrderService)ctx.getBean("orderServiceImpl");
+			service.deleteOrder(orderId);
+			response.getWriter().print("{\"valid\":1}");
+		}else{
+			response.getWriter().print("{\"valid\":0}");
+		}
+		
 	}
+	
+	
+	
 	//修改维修单
 	@ResponseBody
 	@RequestMapping("updateOrder.do")
